refactor: constructionInfo -> spot 컴포넌트로 변경
