+++ packages/mobile/src/page/Map/Spot/style.module.scss
